refactor: abstract functions
